+++ azure-ai-language-textanalytics/src/main/java/com/azure/ai/language/textanalytics/models/SentenceTarget.java
@@ -28,13 +28,13 @@ public final class SentenceTarget {
      * The target offset from the start of the sentence.
      */
     @JsonProperty(value = "offset", required = true)
-    private long offset;
+    private int offset;
 
     /*
      * The length of the target.
      */
     @JsonProperty(value = "length", required = true)
-    private long length;
+    private int length;
 
     /*
      * The target text detected.
@@ -62,8 +62,8 @@ public final class SentenceTarget {
     public SentenceTarget(
             @JsonProperty(value = "sentiment", required = true) TokenSentimentValue sentiment,
             @JsonProperty(value = "confidenceScores", required = true) TargetConfidenceScoreLabel confidenceScores,
-            @JsonProperty(value = "offset", required = true) long offset,
-            @JsonProperty(value = "length", required = true) long length,
+            @JsonProperty(value = "offset", required = true) int offset,
+            @JsonProperty(value = "length", required = true) int length,
             @JsonProperty(value = "text", required = true) String text,
             @JsonProperty(value = "relations", required = true) List<TargetRelation> relations) {
         this.sentiment = sentiment;
@@ -97,7 +97,7 @@ public final class SentenceTarget {
      *
      * @return the offset value.
      */
-    public long getOffset() {
+    public int getOffset() {
         return this.offset;
     }
 
@@ -106,7 +106,7 @@ public final class SentenceTarget {
      *
      * @return the length value.
      */
-    public long getLength() {
+    public int getLength() {
         return this.length;
     }
